- implements Hybrid ELL/COO - implements COOMatrix SPMV on CUDA Device - fix bug in CudaUtils_getBestCudaParameters()

+++ include/BlockGridInfo.h
@@ -7,16 +7,16 @@
 typedef struct BlockGridInfo {
     u_int64_t blockSize;
     u_int64_t gridSize;
-    float utilizationSM;
-    float spread;
-    unsigned int numThread;
-    unsigned int wastedThread;
-    float wastedThreadOverNumThread;
-    float utilization;
-    unsigned long numBlockToFillSM;
-    unsigned long maxThreadPerBlock;
-    unsigned long maxBlockSizePerSM;
-    unsigned long maxThreadPerSM;
+    double utilizationSM;
+    double spread;
+    u_int64_t numThread;
+    u_int64_t wastedThread;
+    double wastedThreadOverNumThread;
+    double utilization;
+    u_int64_t numBlockToFillSM;
+    u_int64_t maxThreadPerBlock;
+    u_int64_t maxBlockSizePerSM;
+    u_int64_t maxThreadPerSM;
 } BlockGridInfo;
 
 void BlockGridInfo_outAsJSON(BlockGridInfo *blockGridInfo, FILE *out);
